# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/parse": {
            "post": {
                "description": "Parse the source using the specified parser",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "parser"
                ],
                "summary": "Parse source",
                "parameters": [
                    {
                        "description": "Parse Source Form",
                        "name": "form",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.ParseSourceForm"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.ParseSourceResponse"
                        }
                    }
                }
            }
        },
        "/scrape": {
            "post": {
                "description": "Scrape the source using the specified scraper",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "scraper"
                ],
                "summary": "Scrape source",
                "parameters": [
                    {
                        "description": "Scrape Form",
                        "name": "form",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.ScrapeForm"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.ScrapeResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "api.ParseSourceForm": {
            "type": "object",
            "required": [
                "parser",
                "source"
            ],
            "properties": {
                "parser": {
                    "type": "string"
                },
                "source": {
                    "type": "string"
                }
            }
        },
        "api.ParseSourceResponse": {
            "type": "object",
            "properties": {
                "files": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "season": {
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "api.ScrapeForm": {
            "type": "object",
            "required": [
                "scraper",
                "title"
            ],
            "properties": {
                "language": {
                    "type": "string"
                },
                "scraper": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "api.ScrapeResponse": {
            "type": "object",
            "properties": {
                "first_air_year": {
                    "type": "string"
                },
                "original_title": {
                    "type": "string"
                },
                "overview": {
                    "type": "string"
                },
                "poster_path": {
                    "type": "string"
                },
                "scraper": {
                    "type": "string"
                },
                "seasons": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/api.ScrapeSeasonResponse"
                    }
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "api.ScrapeSeasonResponse": {
            "type": "object",
            "properties": {
                "overview": {
                    "type": "string"
                },
                "poster": {
                    "type": "string"
                },
                "season_id": {
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                }
            }
        }
    }
}